:rocket: Run frontend towards production backend
Run `npm run dev:prod-api` to target production api
