+++ README.md
@@ -64,13 +64,13 @@ Essentially:
 #### Requirements:
 
 - NodeJS LTS https://nodejs.org/en/
-- Yarn / npm
+- npm / yarn
 
 #### Development
 
-Install dependencies: `yarn` or `npm i`
+Install dependencies: `npm i` or `yarn`
 
-Start the application: `yarn run dev` or `npm run dev`
+Start the application: `npm run dev` or `yarn run dev`
 
 To make sessions acquired from logging in with swagger persist to the frontend, make sure you access them using the same hostname (either 127.0.0.1 or localhost, as long as you use the same both places)
 
@@ -81,6 +81,14 @@ To make sessions acquired from logging in with swagger persist to the frontend, 
 3. `npm run dev`
 4. Open http://localhost:5173/ in browser
 5. login with username: `friByte` and password: `super-duper-secret-password` (password is defined in `friByte.capture-the-flag.service/friByte.capture-the-flag.service/appsettings.json`)
+
+#### Run frontend without backend:
+
+1. `cd frontend`
+2. `npm install`
+3. `npm run dev:prod-api` -> And frontend will target the production backend https://ctf-api.fribyte.no/swagger
+4. Open http://localhost:5173/ in browser
+5. login with username: `friByte` and password would probably be the usual
 
 #### Regenerating react query hooks
