+++ frontend/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "dev:prod-api": "cross-env VITE_APP_API_URL=https://ctf-api.fribyte.no vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "prepare": "cd .. && husky install frontend/.husky",
@@ -32,6 +33,7 @@
     "@types/react-dom": "^18.2.18",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.17",
+    "cross-env": "^7.0.3",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "husky": "^8.0.3",
