+++ frontend/src/config.ts
@@ -1,0 +1,4 @@
+export default {
+  APP_API_URL: import.meta.env.VITE_APP_API_URL,
+  ...window.env,
+};
